refactor(durable): freeze host API, snapshot row, and required stores
Wait-loop glue moves to durable/internal. Temporal activities persist one
Snapshot row. New and NewWorker panic without Catalog, Snapshots, and
Secrets. Prompt/Resume always Put; an empty child bag falls back to the
parent.

Fixes #132
Fixes #133
