+++ durable/internal/doc.go
@@ -1,0 +1,3 @@
+// Package adapter is the Runtime wait-loop glue. Hosts use durable.Runtime,
+// not these functions.
+package adapter
