+++ durable/internal/vfs_test.go
@@ -1,4 +1,4 @@
-package durable
+package adapter
 
 import (
 	"context"
@@ -8,9 +8,8 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/ryanaldo34/tacklr"
 	"github.com/ryanaldo34/tacklr/builtins"
-	"github.com/ryanaldo34/tacklr/internal/testkit"
+	"github.com/ryanaldo34/tacklr/durable"
 	"github.com/ryanaldo34/tacklr/vfs"
 )
 
@@ -20,16 +19,15 @@ func (downProjection) Available() bool                        { return false }
 func (downProjection) Attach(*vfs.MountSession, string) error { return nil }
 
 func TestOpenTurnVFS_nilWhenNoOpenOrProjection(t *testing.T) {
-	ms, err := OpenTurnVFS(t.Context(), "s", AgentSpec{}, nil, vfs.DirectProjection{})
+	ms, err := OpenTurnVFS(t.Context(), "s", durable.AgentSpec{}, nil, vfs.DirectProjection{})
 	if err != nil || ms != nil {
 		t.Fatalf("no OpenVFS: %v %v", ms, err)
 	}
-	ms, err = OpenTurnVFS(t.Context(), "s", AgentSpec{OpenVFS: vfs.Tree(vfs.At("scratch", builtins.Local(t.TempDir())))}, nil, downProjection{})
+	ms, err = OpenTurnVFS(t.Context(), "s", durable.AgentSpec{OpenVFS: vfs.Tree(vfs.At("scratch", builtins.Local(t.TempDir())))}, nil, downProjection{})
 	if err != nil || ms != nil {
 		t.Fatalf("projection down: %v %v", ms, err)
 	}
 	CloseTurnVFS(nil, "s", "test")
-	ClearSessionVFS(nil, "s")
 }
 
 func TestOpenTurnSessions_skillsWithoutProjection(t *testing.T) {
@@ -42,7 +40,7 @@ func TestOpenTurnSessions_skillsWithoutProjection(t *testing.T) {
 	if err := os.WriteFile(filepath.Join(d, "SKILL.md"), []byte("---\nname: research\ndescription: d\n---\n\nbody\n"), 0o644); err != nil {
 		t.Fatal(err)
 	}
-	ws, skills, err := OpenTurnSessions(ctx, "s", AgentSpec{
+	ws, skills, err := OpenTurnSessions(ctx, "s", durable.AgentSpec{
 		OpenVFS:    vfs.Tree(vfs.At("work", builtins.Local(t.TempDir()))),
 		OpenSkills: vfs.Tree(vfs.At("skills", builtins.Local(pack))),
 	}, nil, downProjection{})
@@ -57,7 +55,7 @@ func TestOpenTurnSessions_skillsWithoutProjection(t *testing.T) {
 }
 
 func TestOpenTurnSessions_skillsError(t *testing.T) {
-	_, _, err := OpenTurnSessions(t.Context(), "s", AgentSpec{
+	_, _, err := OpenTurnSessions(t.Context(), "s", durable.AgentSpec{
 		OpenVFS: vfs.Tree(vfs.At("work", builtins.Local(t.TempDir()))),
 		OpenSkills: func(context.Context, string, vfs.Request) (*vfs.MountSession, error) {
 			return nil, os.ErrPermission
@@ -74,7 +72,7 @@ func TestOpenTurnVFS_treeUnderWorkspace(t *testing.T) {
 	if err := os.WriteFile(filepath.Join(dir, "hello.txt"), []byte("hi"), 0o644); err != nil {
 		t.Fatal(err)
 	}
-	ms, err := OpenTurnVFS(ctx, "sess", AgentSpec{
+	ms, err := OpenTurnVFS(ctx, "sess", durable.AgentSpec{
 		OpenVFS: vfs.Tree(vfs.At("docs", builtins.Local(dir))),
 	}, nil, vfs.DirectProjection{})
 	if err != nil || ms == nil {
@@ -90,7 +88,7 @@ func TestOpenTurnVFS_treeUnderWorkspace(t *testing.T) {
 func TestOpenTurnVFS_driveSkippedWithoutToken(t *testing.T) {
 	ctx := t.Context()
 	dir := t.TempDir()
-	ms, err := OpenTurnVFS(ctx, "s", AgentSpec{
+	ms, err := OpenTurnVFS(ctx, "s", durable.AgentSpec{
 		OpenVFS: vfs.Tree(vfs.At("scratch", builtins.Local(dir))),
 	}, nil, vfs.DirectProjection{})
 	if err != nil || ms == nil {
@@ -108,7 +106,7 @@ func (failAttach) Available() bool                          { return true }
 func (f failAttach) Attach(*vfs.MountSession, string) error { return f.err }
 
 func TestOpenTurnVFS_attachError(t *testing.T) {
-	_, err := OpenTurnVFS(t.Context(), "s", AgentSpec{
+	_, err := OpenTurnVFS(t.Context(), "s", durable.AgentSpec{
 		OpenVFS: vfs.Tree(vfs.At("scratch", builtins.Local(t.TempDir()))),
 	}, nil, failAttach{err: os.ErrPermission})
 	if err == nil {
@@ -117,40 +115,10 @@ func TestOpenTurnVFS_attachError(t *testing.T) {
 }
 
 func TestOpenTurnVFS_unknownAtName(t *testing.T) {
-	_, err := OpenTurnVFS(t.Context(), "s", AgentSpec{
+	_, err := OpenTurnVFS(t.Context(), "s", durable.AgentSpec{
 		OpenVFS: vfs.Tree(vfs.At("", builtins.Local(t.TempDir()))),
 	}, nil, vfs.DirectProjection{})
 	if err == nil {
 		t.Fatal("want At name error")
-	}
-}
-
-func TestClearSessionVFS_noop(t *testing.T) {
-	cat := NewCatalog("default")
-	cat.Register("default", AgentSpec{
-		Options: tacklr.AgentOptions{Model: &testkit.ScriptedModel{}, Config: tacklr.Config{MaxWindowSize: 8192}},
-		OpenVFS: vfs.Tree(vfs.At("scratch", builtins.Local(t.TempDir()))),
-	})
-	ClearSessionVFS(cat, "s1")
-}
-
-func TestBindingsForTurn_skipsEmptyAliasAndEmptyToken(t *testing.T) {
-	recipes := []MountRecipe{
-		{Provider: "gdrive", Alias: ""},
-		{Provider: "gdrive", Alias: "docs", Params: nil},
-	}
-	got := BindingsForTurn(recipes, AuthContext{Bindings: []vfs.Binding{{
-		Provider: "gdrive",
-		Auth:     vfs.Credential{Token: "   "},
-	}}})
-	if len(got) != 0 {
-		t.Fatalf("want none, got %+v", got)
-	}
-	got = BindingsForTurn(recipes, AuthContext{Bindings: []vfs.Binding{{
-		Provider: "gdrive",
-		Auth:     vfs.Credential{Token: "tok"},
-	}}})
-	if len(got) != 1 || got[0].Params[vfs.ParamName] != "docs" {
-		t.Fatalf("token-only provider: %+v", got)
 	}
 }
