+++ durable/temporal/child.go
@@ -10,6 +10,7 @@ import (
 
 	"github.com/ryanaldo34/tacklr"
 	"github.com/ryanaldo34/tacklr/durable"
+	adapter "github.com/ryanaldo34/tacklr/durable/internal"
 	"github.com/ryanaldo34/tacklr/interrupt"
 )
 
@@ -72,7 +73,7 @@ func spawnedNudge(spawned []childRun) string {
 	for i, c := range spawned {
 		rows[i] = durable.SessionStatus{ID: c.id, Specialist: c.spec, State: durable.SessionRunning}
 	}
-	return durable.ChildrenNudge(rows)
+	return adapter.ChildrenNudge(rows)
 }
 
 func cancelOne(ctx workflow.Context, spawned *[]childRun, id durable.SessionID) {
@@ -197,13 +198,13 @@ type activityChildren struct {
 }
 
 func (a *activityChildren) SpawnChild(_ context.Context, specialist, task, callID string) (string, error) {
-	specialist, task, err := durable.NormalizeSpawn(specialist, task)
+	specialist, task, err := adapter.NormalizeSpawn(specialist, task)
 	if err != nil {
 		return "", err
 	}
 	if a.catalog != nil {
 		if spec, ok := a.catalog.Lookup(a.agentID); ok {
-			if _, err := durable.OverlaySpecialist(spec, specialist); err != nil {
+			if _, err := adapter.OverlaySpecialist(spec, specialist); err != nil {
 				return "", fmt.Errorf("%w: %w", tacklr.ErrNotFound, err)
 			}
 		}
@@ -237,7 +238,7 @@ func (a *activityChildren) Children() []tacklr.Child {
 func (a *activityChildren) CancelChild(_ context.Context, id string) error {
 	sid := durable.SessionID(id)
 	if sid != a.spawnID && !slices.Contains(a.known, sid) {
-		return durable.UnknownChild(id)
+		return adapter.UnknownChild(id)
 	}
 	a.cancelID = sid
 	return nil
@@ -246,7 +247,7 @@ func (a *activityChildren) CancelChild(_ context.Context, id string) error {
 func (a *activityChildren) AwaitChild(_ context.Context, id, _ string) (tacklr.Child, error) {
 	sid := durable.SessionID(id)
 	if sid != a.spawnID && !slices.Contains(a.known, sid) {
-		return tacklr.Child{}, durable.UnknownChild(id)
+		return tacklr.Child{}, adapter.UnknownChild(id)
 	}
 	a.awaitID = sid
 	return tacklr.Child{}, &interrupt.ChildWaiting{Kind: interrupt.TypeChildWaiting}
