+++ durable/temporal/runtime.go
@@ -1,6 +1,7 @@
 // Package temporal is the Temporal adapter for durable.Runtime.
-// Hosts use Dial, New, and NewWorker with the same Config (including Secrets).
-// NewWorker registers SessionWorkflow and the turn activities.
+// Hosts use Dial, New, and NewWorker with the same Config (including
+// Snapshots and Secrets). NewWorker registers SessionWorkflow and the
+// turn activities.
 package temporal
 
 import (
@@ -18,6 +19,7 @@ import (
 	"github.com/ryanaldo34/tacklr"
 	"github.com/ryanaldo34/tacklr/durable"
 	"github.com/ryanaldo34/tacklr/durable/inprocess"
+	adapter "github.com/ryanaldo34/tacklr/durable/internal"
 	"github.com/ryanaldo34/tacklr/mcp"
 	"github.com/ryanaldo34/tacklr/vfs"
 )
@@ -69,8 +71,10 @@ func resolveActivityAttempts(n int32) int32 {
 
 // Config is the single Temporal host config for New and NewWorker.
 type Config struct {
-	Catalog    durable.Catalog
-	TaskQueue  string
+	Catalog   durable.Catalog
+	TaskQueue string
+	// Snapshots is the session record. Required. New and NewWorker must share
+	// the same instance. Tokens never go here.
 	Snapshots  durable.SnapshotStore
 	Fallback   durable.EventLog
 	Projection vfs.Projection
@@ -97,11 +101,16 @@ func (c Config) queue() string {
 	return c.TaskQueue
 }
 
-func (c Config) snaps() durable.SnapshotStore {
-	if c.Snapshots != nil {
-		return c.Snapshots
+func requireCfg(cfg Config) {
+	if cfg.Catalog == nil {
+		panic("temporal: Catalog is required")
 	}
-	return inprocess.NewMemorySnapshot()
+	if cfg.Snapshots == nil {
+		panic("temporal: Snapshots is required")
+	}
+	if cfg.Secrets == nil {
+		panic("temporal: Secrets is required")
+	}
 }
 
 func (c Config) memoryLog() *inprocess.MemoryEventLog {
@@ -112,23 +121,18 @@ func (c Config) memoryLog() *inprocess.MemoryEventLog {
 }
 
 // New constructs a Temporal Runtime. The host must also run NewWorker on the
-// same Config, including the same Secrets store.
+// same Config, including the same Snapshots and Secrets stores.
 func New(c client.Client, cfg Config) *Runtime {
 	if c == nil {
 		panic("temporal: Client is required")
 	}
-	if cfg.Catalog == nil {
-		panic("temporal: Catalog is required")
-	}
-	if cfg.Secrets == nil {
-		panic("temporal: Secrets is required")
-	}
+	requireCfg(cfg)
 	return &Runtime{
 		client:              c,
 		taskQueue:           cfg.queue(),
 		catalog:             cfg.Catalog,
 		fallback:            cfg.memoryLog(),
-		snapshots:           cfg.snaps(),
+		snapshots:           cfg.Snapshots,
 		disableStreams:      cfg.DisableStreams,
 		turnLocalityTimeout: cfg.TurnLocality,
 		activityTimeout:     resolveActivityTimeout(cfg.ActivityTimeout),
@@ -152,7 +156,7 @@ func (r *Runtime) CreateSession(ctx context.Context, req durable.CreateSession) 
 	if id == "" {
 		id = durable.SessionID(uuid.NewString())
 	}
-	seed, err := durable.EncodeUserState(req.State)
+	seed, err := adapter.EncodeUserState(req.State)
 	if err != nil {
 		return "", err
 	}
@@ -200,14 +204,12 @@ func (r *Runtime) signal(ctx context.Context, id durable.SessionID, name string,
 }
 
 func (r *Runtime) Prompt(ctx context.Context, sessionID durable.SessionID, msg durable.Prompt) error {
-	encoded, err := durable.EncodeUserState(msg.State)
+	encoded, err := adapter.EncodeUserState(msg.State)
 	if err != nil {
 		return err
 	}
-	if len(msg.Auth.Bindings) > 0 {
-		if err := r.secrets.Put(ctx, sessionID, durable.Secrets{Auth: msg.Auth}); err != nil {
-			return err
-		}
+	if err := r.secrets.Put(ctx, sessionID, durable.Secrets{Auth: msg.Auth}); err != nil {
+		return err
 	}
 	return r.signal(ctx, sessionID, signalPrompt, promptSignal{
 		Text:        msg.Text,
@@ -220,14 +222,12 @@ func (r *Runtime) Prompt(ctx context.Context, sessionID durable.SessionID, msg d
 }
 
 func (r *Runtime) Resume(ctx context.Context, sessionID durable.SessionID, resume durable.Resume) error {
-	encoded, err := durable.EncodeUserState(resume.State)
+	encoded, err := adapter.EncodeUserState(resume.State)
 	if err != nil {
 		return err
 	}
-	if len(resume.Auth.Bindings) > 0 {
-		if err := r.secrets.Put(ctx, sessionID, durable.Secrets{Auth: resume.Auth}); err != nil {
-			return err
-		}
+	if err := r.secrets.Put(ctx, sessionID, durable.Secrets{Auth: resume.Auth}); err != nil {
+		return err
 	}
 	return r.signal(ctx, sessionID, signalResume, resumeSignal{
 		Responses: resume.Responses,
@@ -259,7 +259,6 @@ func (r *Runtime) Close(ctx context.Context, sessionID durable.SessionID) error 
 	_ = r.secrets.Delete(ctx, sessionID)
 	_ = r.snapshots.Delete(ctx, sessionID)
 	_ = r.fallback.CloseSession(ctx, sessionID)
-	durable.ClearSessionVFS(r.catalog, sessionID)
 	return nil
 }
 
@@ -374,5 +373,3 @@ func (r *Runtime) Status(ctx context.Context, id durable.SessionID) (durable.Ses
 	_ = val.Get(&st)
 	return st, nil
 }
-
-var _ durable.Runtime = (*Runtime)(nil)
